Annotate gene for intervals (CNV, CNVR, ROH or QTL)

This function is used for annotating genes over any interval by giving the interval list and reference gene list. The interval file requires at least four columns: Interval_ID, Chr, Start and End. The first column must be the ID column, the column of Chr should only contain the chromosome number (i.e., no "chr" prefix), and the units of the Start and End columns are the basepair.

call_gene(
  refgene = "ARS_ens",
  interval = NULL,
  clean_cnv = NULL,
  folder = "call_gene"
)

Arguments

refgene
a file containing the gene information. This should be prepared using the [get_refgene()] function.

interval
a file containing a list of genomic intervals, such as CNVs, ROHs, or QTLs. The file must contain at least the following four columns: Interval_ID, Chr, Start and End

clean_cnv
the output data from a previous run of the cnv_clean function. If this is provided, additional output files are created that annotate the CNVs

folder
set the name of the output folder

Value

Annotated file and brief summary