Custom visualizing CNV

This function is designed to custom plot CNVs, it now support to visualize CNVs by Chromosome, Interested Individual, Interested Region, Target Gene, et al.

cnv_visual(
  clean_cnv,
  max_chr = NULL,
  chr_id = NULL,
  species = NULL,
  start_position = NULL,
  end_position = NULL,
  individual_id = NULL,
  target_gene = NULL,
  max_chr_length = 160,
  refgene = NULL,
  plot_title = NULL,
  report_id = NULL,
  pedigree = NULL,
  show_name = c(0, 160),
  width_1 = 13,
  height_1 = 10,
  folder = "cnv_visual",
  col_0 = "hotpink",
  col_1 = "turquoise",
  col_2 = "gray",
  col_3 = "tomato",
  col_4 = "deepskyblue",
  col_gene = "gray",
  gene_font_size = 2.2
)

Arguments

clean_cnv
cnv list file, standard file was generated by 'cnv_clean' function

max_chr
the maximum number of chromosomes to plot, it used for plot all chromosomes at once

chr_id
the number of chromosome want to plot

species
which species in the input data, unused at the moment

start_position
decimal digit, default unit is 'Mb'. such as 23.2112

end_position
decimal digit, default unit is 'Mb'. such as 23.2112

individual_id
the ID of individual in cnv list, used to plot all chromosome for specific Individual

target_gene
the

max_chr_length
the maximum length of chromosome, default unit is 'Mb'

refgene
if true, will plot gene above the CNV plot, need to combine with the clean_cnv, and the stardard clean_cnv file need the annotated cnv list which was generated from call_gene function, internal reference gene-list are "ARS_ens", "ARS_UCSC" and "UMD_UCSC". Or provide the reference genes list corresponding to your data

plot_title
set the title of final plot

report_id
report the sample ID while plotting

pedigree
pedigree list, require at least three columns, Sample_Id, Sire and Dam

show_name
default value is show_name = c(0, 160). accept the vectors only, unit is Mb. for example show_name = c(11.2, 12.4, 15.3, 18.4), means only plot the genes within the given interval

width_1
number to set the width of final plot size, unit is 'cm'

height_1
number to set the height of final plot size, unit is 'cm'

folder
set the name of folder to save results

col_0
set color for 0 copy of CNV

col_1
set color for 1 copy of CNV

col_2
set color for 2 copy of CNV (which might be ROH)

col_3
set color for 3 copy of CNV

col_4
set color for 4 copy of CNV

col_gene
set the color of Gene, only work in Gene annotated at CNV plot

gene_font_size
set the size of Gene in CNV Annotation plot

Value

CNV distribution plot