Custom CNVR distribution map and plot all high frequency CNVR at once

Custom CNVR distribution map and plot all high frequency CNVR at once

cnvr_plot(
  cnvr,
  assembly = "ARS",
  legend_x = 127,
  legend_y = 30,
  clean_cnv = NULL,
  sample_size = NULL,
  common_cnv_threshold = 0.05,
  refgene = "ARS",
  gain_col = "red",
  loss_col = "green",
  mixed_col = "blue",
  folder = "cnvr_plot",
  gene_font_size = 2.2,
  width_1 = 14,
  height_1 = 10,
  col_gene = "gray"
)

Arguments

cnvr
standard cnvr file was generated by 'call_cnvr' function

assembly
which reference genome assembly used in the data, each chromosome has different length between various assembly, now only support Bovine UMD3.1 and ARS1.2

legend_x
the x coordinate of legend, relative to the maximum length of Chromosome, unit is 'Mb'

legend_y
the y coordinate of legend

clean_cnv
standard input file was generated by 'cnv_clean' function

sample_size
integer, the total number of unique samples in the cnv result. combine with common_cnv_threshold to plot all CNVs which passed the threshold

common_cnv_threshold
two decimal places, combine with sample_size to plot all CNVs passed the common threshold

refgene
reference gene list, use for plot gene

gain_col
set color for type of gain CNVR

loss_col
set color for type of loss CNVR

mixed_col
set color for type of mixed CNVR

folder
set name of folder to save results

gene_font_size
set the font size of gene in CNV annotation plot

width_1
number to set the width of final plot size, unit is 'cm'

height_1
number to set the height of final plot size, unit is 'cm'

col_gene
set the color of Gene, only work in Gene annotated at CNV plot

Value

A figure of CNVR distribution map and plot parameters. If given clean_cnv file, will plot all CNVRs which are passed common threshold.