Find the consensus genes by comparing the frequency of gene in which has CNV status

Do comparison of genes between multiple annotated gene frequency lists. The default genelist file was generate by call_gene function. If the users customize the genelists, it requires at least two columns in each gene list, including fixed column names name2(Gene name) and Frequency(frequency of gene)

compare_gene(
  gene_freq_1,
  gene_freq_2,
  gene_freq_3 = NULL,
  gene_freq_4 = NULL,
  common_gene_threshold = 3,
  title_1 = "list_1",
  title_2 = "list_2",
  title_3 = "list_3",
  title_4 = "list_4",
  height_1 = 10,
  width_1 = 14,
  folder = "compre_gene",
  col_1 = "red",
  col_2 = "yellow",
  col_3 = "dodgerblue",
  col_4 = "pink2",
  color_label = FALSE
)

Arguments

gene_freq_1
first genelist, the default file was generated from call_gene function

gene_freq_2
second genelist, the default file was generated from call_gene function

gene_freq_3
third genelist, the default file was generated from call_gene function

gene_freq_4
fourth genelist, the default file was generated from call_gene function

common_gene_threshold
integer input.could calculate by multiplying The Number of Sample by 0.05 or 0.1, et al

title_1
set title(two and three lists) or label(four lists) in figure

title_2
set title(two and three lists) or label(four lists) in figure

title_3
set title(two and three lists) or label(four lists) in figure

title_4
set title(two and three lists) or label(four lists) in figure

height_1
customize the height to save figure

width_1
customize the width to save figure

folder
set the name of folder to save results

col_1
set color for common_high in two gene lists comparison plot or the color of High frequency gene in multiple comparison heatmap plot

col_2
set color for common_low in two gene lists comparison plot or the color of Low frequency gene in multiple comparison heatmap plot

col_3
set color for High_Freq_list_1 in two gene lists comparison plot

col_4
set color for High_Freq_list_2 in two gene lists comparison plot

color_label
if TRUE, the color of gene labels will marked in heatmap, the red labels are genes that passed common threshold

Value

Comparison summary results and plots.