Convert map file

Prepare map files for PennCNV and Plink with Default and Target assembly. Four columns in fixed order are required for both default and target map file, they are Chr, Name, Morgan, Position

convert_map(
  default_map,
  target_map,
  chr_set = 29,
  defMap_title = "Default",
  tarMap_title = "Target",
  species = "Bovine",
  col_1 = "green4",
  col_2 = "red1",
  col_3 = "deeppink2",
  col_4 = "deeppink2",
  col_5 = "turquoise3",
  col_6 = "turquoise3",
  legend_position = c(0.9, 0.9),
  height_c = 16,
  width_c = 22,
  height_d = 15,
  width_d = 25
)

Arguments

default_map
the map file to be converted

target_map
the map file used to convert

chr_set
set the number of autosomes

defMap_title
customize the title of default map in plot

tarMap_title
customize the title of target map in plot

species
the name of species

col_1
set color for the type of Match in SNP comparison plot

col_2
set color for the type of Mismatch in SNP comparison plot

col_3
set color for the bar of Target_Map in SNP density plot

col_4
set color for line of Target_map in density plot

col_5
set color for point of Default_Map in SNP density plot

col_6
set color for line of Default_map in SNP density plot

legend_position
default value = c(0.9, 0.9), used to adjust the legend in map comparison plot

height_c
set height for Comparison plot

width_c
set width for comparison plot

height_d
set height for SNP density plot

width_d
set width for SNP density plot

Value

Details of comparison results between given map, and standard input map files used in PennCNV and Plink