Plotting CNVR with all relevant information

Plotting CNVR with all relevant information

plot_cnvr_panorama(
  cnvr,
  cnv_annotation,
  intensity = NULL,
  map = NULL,
  prefix_bed = NULL,
  ld_heat = TRUE,
  sample_size,
  common_cnv_threshold = 0.05,
  width_1 = 14,
  height_1 = 30,
  folder = "cnvr_panorama",
  col_0 = "hotpink",
  col_1 = "turquoise",
  col_2 = "NA",
  col_3 = "tomato",
  col_4 = "deepskyblue",
  col_gene = "black",
  gene_font_size = 2.2
)

Arguments

cnvr
cnvr list, standard file generated by the call_cnvr function

cnv_annotation
standard file generated by the call_gene function

intensity
the signal intensity file contains LRR and BAF

map
the map corresponding to the reference genome used to create the CNVR file, standard file generated by the convert_map function

prefix_bed
the prefix of the BED, BIM and FAM files in Plink format

ld_heat
If TRUE, will plot a heatmap of LD, otherwise will plot the classic inverted triangle plot. Plotting is faster with TRUE, and plotting the triange with FALSE may fail for large regions.

sample_size
the total number of unique samples in the CNVR list

common_cnv_threshold
two decimal places, combine with sample_size to set the common threshold

width_1
default value is 14, unit is 'cm', set the width of final the plot

height_1
default value is 30, unit is 'cm', set the height of final the plot

folder
set the folder name to save results in

col_0
set colour for 0 copy of CNV

col_1
set colour for 1 copy of CNV

col_2
set colour for 2 copy of CNV (which might be ROH)

col_3
set colour for 3 copy of CNV

col_4
set colour for 4 copy of CNV

col_gene
set colour of genes

gene_font_size
set the font size of gene labels in CNV annotated plot

Value

cnvr plot with all CNVs, annotated genes, Log R Ratio, B Allele Frequency, Genotyping rate and LD...